Remove parser from source and revert exports
