+++ js/dcr_parser.js
@@ -1,2170 +1,2068 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.parser = void 0;
-const dcr_1 = require("./dcr");
-let parser = /*
+parser = /*
  * Generated by PEG.js 0.10.0.
  *
  * http://pegjs.org/
- */ (function () {
-    "use strict";
-    function peg$subclass(child, parent) {
-        function ctor() { this.constructor = child; }
-        ctor.prototype = parent.prototype;
-        child.prototype = new ctor();
+ */
+(function() {
+  "use strict";
+
+  function peg$subclass(child, parent) {
+    function ctor() { this.constructor = child; }
+    ctor.prototype = parent.prototype;
+    child.prototype = new ctor();
+  }
+
+  function peg$SyntaxError(message, expected, found, location) {
+    this.message  = message;
+    this.expected = expected;
+    this.found    = found;
+    this.location = location;
+    this.name     = "SyntaxError";
+
+    if (typeof Error.captureStackTrace === "function") {
+      Error.captureStackTrace(this, peg$SyntaxError);
     }
-    function peg$SyntaxError(message, expected, found, location) {
-        this.message = message;
-        this.expected = expected;
-        this.found = found;
-        this.location = location;
-        this.name = "SyntaxError";
-        if (typeof Error.captureStackTrace === "function") {
-            Error.captureStackTrace(this, peg$SyntaxError);
-        }
-    }
-    peg$subclass(peg$SyntaxError, Error);
-    peg$SyntaxError.buildMessage = function (expected, found) {
-        var DESCRIBE_EXPECTATION_FNS = {
-            literal: function (expectation) {
-                return "\"" + literalEscape(expectation.text) + "\"";
-            },
-            "class": function (expectation) {
-                var escapedParts = "", i;
-                for (i = 0; i < expectation.parts.length; i++) {
-                    escapedParts += expectation.parts[i] instanceof Array
-                        ? classEscape(expectation.parts[i][0]) + "-" + classEscape(expectation.parts[i][1])
-                        : classEscape(expectation.parts[i]);
-                }
-                return "[" + (expectation.inverted ? "^" : "") + escapedParts + "]";
-            },
-            any: function (expectation) {
-                return "any character";
-            },
-            end: function (expectation) {
-                return "end of input";
-            },
-            other: function (expectation) {
-                return expectation.description;
+  }
+
+  peg$subclass(peg$SyntaxError, Error);
+
+  peg$SyntaxError.buildMessage = function(expected, found) {
+    var DESCRIBE_EXPECTATION_FNS = {
+          literal: function(expectation) {
+            return "\"" + literalEscape(expectation.text) + "\"";
+          },
+
+          "class": function(expectation) {
+            var escapedParts = "",
+                i;
+
+            for (i = 0; i < expectation.parts.length; i++) {
+              escapedParts += expectation.parts[i] instanceof Array
+                ? classEscape(expectation.parts[i][0]) + "-" + classEscape(expectation.parts[i][1])
+                : classEscape(expectation.parts[i]);
             }
+
+            return "[" + (expectation.inverted ? "^" : "") + escapedParts + "]";
+          },
+
+          any: function(expectation) {
+            return "any character";
+          },
+
+          end: function(expectation) {
+            return "end of input";
+          },
+
+          other: function(expectation) {
+            return expectation.description;
+          }
         };
-        function hex(ch) {
-            return ch.charCodeAt(0).toString(16).toUpperCase();
+
+    function hex(ch) {
+      return ch.charCodeAt(0).toString(16).toUpperCase();
+    }
+
+    function literalEscape(s) {
+      return s
+        .replace(/\\/g, '\\\\')
+        .replace(/"/g,  '\\"')
+        .replace(/\0/g, '\\0')
+        .replace(/\t/g, '\\t')
+        .replace(/\n/g, '\\n')
+        .replace(/\r/g, '\\r')
+        .replace(/[\x00-\x0F]/g,          function(ch) { return '\\x0' + hex(ch); })
+        .replace(/[\x10-\x1F\x7F-\x9F]/g, function(ch) { return '\\x'  + hex(ch); });
+    }
+
+    function classEscape(s) {
+      return s
+        .replace(/\\/g, '\\\\')
+        .replace(/\]/g, '\\]')
+        .replace(/\^/g, '\\^')
+        .replace(/-/g,  '\\-')
+        .replace(/\0/g, '\\0')
+        .replace(/\t/g, '\\t')
+        .replace(/\n/g, '\\n')
+        .replace(/\r/g, '\\r')
+        .replace(/[\x00-\x0F]/g,          function(ch) { return '\\x0' + hex(ch); })
+        .replace(/[\x10-\x1F\x7F-\x9F]/g, function(ch) { return '\\x'  + hex(ch); });
+    }
+
+    function describeExpectation(expectation) {
+      return DESCRIBE_EXPECTATION_FNS[expectation.type](expectation);
+    }
+
+    function describeExpected(expected) {
+      var descriptions = new Array(expected.length),
+          i, j;
+
+      for (i = 0; i < expected.length; i++) {
+        descriptions[i] = describeExpectation(expected[i]);
+      }
+
+      descriptions.sort();
+
+      if (descriptions.length > 0) {
+        for (i = 1, j = 1; i < descriptions.length; i++) {
+          if (descriptions[i - 1] !== descriptions[i]) {
+            descriptions[j] = descriptions[i];
+            j++;
+          }
         }
-        function literalEscape(s) {
-            return s
-                .replace(/\\/g, '\\\\')
-                .replace(/"/g, '\\"')
-                .replace(/\0/g, '\\0')
-                .replace(/\t/g, '\\t')
-                .replace(/\n/g, '\\n')
-                .replace(/\r/g, '\\r')
-                .replace(/[\x00-\x0F]/g, function (ch) { return '\\x0' + hex(ch); })
-                .replace(/[\x10-\x1F\x7F-\x9F]/g, function (ch) { return '\\x' + hex(ch); });
+        descriptions.length = j;
+      }
+
+      switch (descriptions.length) {
+        case 1:
+          return descriptions[0];
+
+        case 2:
+          return descriptions[0] + " or " + descriptions[1];
+
+        default:
+          return descriptions.slice(0, -1).join(", ")
+            + ", or "
+            + descriptions[descriptions.length - 1];
+      }
+    }
+
+    function describeFound(found) {
+      return found ? "\"" + literalEscape(found) + "\"" : "end of input";
+    }
+
+    return "Expected " + describeExpected(expected) + " but " + describeFound(found) + " found.";
+  };
+
+  function peg$parse(input, options) {
+    options = options !== void 0 ? options : {};
+
+    var peg$FAILED = {},
+
+        peg$startRuleFunctions = { File: peg$parseFile },
+        peg$startRuleFunction  = peg$parseFile,
+
+        peg$c0 = function(g) {return g},
+        peg$c1 = function() {graphs.push(new DCRGraph(top()))},
+        peg$c2 = function() {return graphs.pop()},
+        peg$c3 = "{",
+        peg$c4 = peg$literalExpectation("{", false),
+        peg$c5 = "}",
+        peg$c6 = peg$literalExpectation("}", false),
+        peg$c7 = function(n, l, m, g) {top().addEvent(n, l, m, g)},
+        peg$c8 = function(n, m, g) {top().addEvent(n, n, m, g)},
+        peg$c9 = function(n, l, m) {top().addEvent(n, l, m)},
+        peg$c10 = function(n, m) {top().addEvent(n, n, m )},
+        peg$c11 = "()",
+        peg$c12 = peg$literalExpectation("()", false),
+        peg$c13 = function() {return {ex: false, in: true, pe: false}},
+        peg$c14 = "(",
+        peg$c15 = peg$literalExpectation("(", false),
+        peg$c16 = ")",
+        peg$c17 = peg$literalExpectation(")", false),
+        peg$c18 = function(b1, b2, b3) {return {ex: b1, in: b2, pe: b3}},
+        peg$c19 = "-",
+        peg$c20 = peg$literalExpectation("-", false),
+        peg$c21 = "->*",
+        peg$c22 = peg$literalExpectation("->*", false),
+        peg$c23 = function(es1, d, g, es2) {for (var e1 of es1) for (var e2 of es2) top().addCondition(e1, e2, d, g);},
+        peg$c24 = "-><>",
+        peg$c25 = peg$literalExpectation("-><>", false),
+        peg$c26 = function(es1, g, es2) {for (var e1 of es1) for (var e2 of es2) top().addMilestone(e1, e2, g);},
+        peg$c27 = "*-",
+        peg$c28 = peg$literalExpectation("*-", false),
+        peg$c29 = "->",
+        peg$c30 = peg$literalExpectation("->", false),
+        peg$c31 = function(es1, d, g, es2) {for (var e1 of es1) for (var e2 of es2) top().addResponse(e1, e2, d, g);},
+        peg$c32 = "->+",
+        peg$c33 = peg$literalExpectation("->+", false),
+        peg$c34 = function(es1, g, es2) {for (var e1 of es1) for (var e2 of es2) top().addInclude(e1, e2, g);},
+        peg$c35 = "->%",
+        peg$c36 = peg$literalExpectation("->%", false),
+        peg$c37 = function(es1, g, es2) {for (var e1 of es1) for (var e2 of es2) top().addExclude(e1, e2, g);},
+        peg$c38 = function(i) {return i;},
+        peg$c39 = function() {return undefined;},
+        peg$c40 = "[",
+        peg$c41 = peg$literalExpectation("[", false),
+        peg$c42 = "]",
+        peg$c43 = peg$literalExpectation("]", false),
+        peg$c44 = function(exp) {return exp;},
+        peg$c45 = function() {return true;},
+        peg$c46 = "",
+        peg$c47 = function(n) {return [n]},
+        peg$c48 = function(ns) {return ns},
+        peg$c49 = function(n, ns) {ns.push(n); return ns},
+        peg$c50 = " ",
+        peg$c51 = peg$literalExpectation(" ", false),
+        peg$c52 = function(t, ws, n) {return t.concat(ws.join(''), n)},
+        peg$c53 = /^[a-z]/,
+        peg$c54 = peg$classExpectation([["a", "z"]], false, false),
+        peg$c55 = /^[A-Z]/,
+        peg$c56 = peg$classExpectation([["A", "Z"]], false, false),
+        peg$c57 = /^[0-9]/,
+        peg$c58 = peg$classExpectation([["0", "9"]], false, false),
+        peg$c59 = "_",
+        peg$c60 = peg$literalExpectation("_", false),
+        peg$c61 = function(str) {return str.join('')},
+        peg$c62 = "<",
+        peg$c63 = peg$literalExpectation("<", false),
+        peg$c64 = ">",
+        peg$c65 = peg$literalExpectation(">", false),
+        peg$c66 = function(lab) {return (lab.join(''))},
+        peg$c67 = "0",
+        peg$c68 = peg$literalExpectation("0", false),
+        peg$c69 = function() {return false},
+        peg$c70 = "1",
+        peg$c71 = peg$literalExpectation("1", false),
+        peg$c72 = function() {return true},
+        peg$c73 = "false",
+        peg$c74 = peg$literalExpectation("false", false),
+        peg$c75 = "true",
+        peg$c76 = peg$literalExpectation("true", false),
+        peg$c77 = "f",
+        peg$c78 = peg$literalExpectation("f", false),
+        peg$c79 = "t",
+        peg$c80 = peg$literalExpectation("t", false),
+        peg$c81 = peg$otherExpectation("Integer"),
+        peg$c82 = function() { return parseInt(text(), 10); },
+        peg$c83 = ";",
+        peg$c84 = peg$literalExpectation(";", false),
+        peg$c85 = "\r\n",
+        peg$c86 = peg$literalExpectation("\r\n", false),
+        peg$c87 = "\n",
+        peg$c88 = peg$literalExpectation("\n", false),
+        peg$c89 = peg$anyExpectation(),
+        peg$c90 = ",",
+        peg$c91 = peg$literalExpectation(",", false),
+        peg$c92 = "\t",
+        peg$c93 = peg$literalExpectation("\t", false),
+
+        peg$currPos          = 0,
+        peg$savedPos         = 0,
+        peg$posDetailsCache  = [{ line: 1, column: 1 }],
+        peg$maxFailPos       = 0,
+        peg$maxFailExpected  = [],
+        peg$silentFails      = 0,
+
+        peg$result;
+
+    if ("startRule" in options) {
+      if (!(options.startRule in peg$startRuleFunctions)) {
+        throw new Error("Can't start parsing from rule \"" + options.startRule + "\".");
+      }
+
+      peg$startRuleFunction = peg$startRuleFunctions[options.startRule];
+    }
+
+    function text() {
+      return input.substring(peg$savedPos, peg$currPos);
+    }
+
+    function location() {
+      return peg$computeLocation(peg$savedPos, peg$currPos);
+    }
+
+    function expected(description, location) {
+      location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos)
+
+      throw peg$buildStructuredError(
+        [peg$otherExpectation(description)],
+        input.substring(peg$savedPos, peg$currPos),
+        location
+      );
+    }
+
+    function error(message, location) {
+      location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos)
+
+      throw peg$buildSimpleError(message, location);
+    }
+
+    function peg$literalExpectation(text, ignoreCase) {
+      return { type: "literal", text: text, ignoreCase: ignoreCase };
+    }
+
+    function peg$classExpectation(parts, inverted, ignoreCase) {
+      return { type: "class", parts: parts, inverted: inverted, ignoreCase: ignoreCase };
+    }
+
+    function peg$anyExpectation() {
+      return { type: "any" };
+    }
+
+    function peg$endExpectation() {
+      return { type: "end" };
+    }
+
+    function peg$otherExpectation(description) {
+      return { type: "other", description: description };
+    }
+
+    function peg$computePosDetails(pos) {
+      var details = peg$posDetailsCache[pos], p;
+
+      if (details) {
+        return details;
+      } else {
+        p = pos - 1;
+        while (!peg$posDetailsCache[p]) {
+          p--;
         }
-        function classEscape(s) {
-            return s
-                .replace(/\\/g, '\\\\')
-                .replace(/\]/g, '\\]')
-                .replace(/\^/g, '\\^')
-                .replace(/-/g, '\\-')
-                .replace(/\0/g, '\\0')
-                .replace(/\t/g, '\\t')
-                .replace(/\n/g, '\\n')
-                .replace(/\r/g, '\\r')
-                .replace(/[\x00-\x0F]/g, function (ch) { return '\\x0' + hex(ch); })
-                .replace(/[\x10-\x1F\x7F-\x9F]/g, function (ch) { return '\\x' + hex(ch); });
+
+        details = peg$posDetailsCache[p];
+        details = {
+          line:   details.line,
+          column: details.column
+        };
+
+        while (p < pos) {
+          if (input.charCodeAt(p) === 10) {
+            details.line++;
+            details.column = 1;
+          } else {
+            details.column++;
+          }
+
+          p++;
         }
-        function describeExpectation(expectation) {
-            return DESCRIBE_EXPECTATION_FNS[expectation.type](expectation);
+
+        peg$posDetailsCache[pos] = details;
+        return details;
+      }
+    }
+
+    function peg$computeLocation(startPos, endPos) {
+      var startPosDetails = peg$computePosDetails(startPos),
+          endPosDetails   = peg$computePosDetails(endPos);
+
+      return {
+        start: {
+          offset: startPos,
+          line:   startPosDetails.line,
+          column: startPosDetails.column
+        },
+        end: {
+          offset: endPos,
+          line:   endPosDetails.line,
+          column: endPosDetails.column
         }
-        function describeExpected(expected) {
-            var descriptions = new Array(expected.length), i, j;
-            for (i = 0; i < expected.length; i++) {
-                descriptions[i] = describeExpectation(expected[i]);
-            }
-            descriptions.sort();
-            if (descriptions.length > 0) {
-                for (i = 1, j = 1; i < descriptions.length; i++) {
-                    if (descriptions[i - 1] !== descriptions[i]) {
-                        descriptions[j] = descriptions[i];
-                        j++;
-                    }
+      };
+    }
+
+    function peg$fail(expected) {
+      if (peg$currPos < peg$maxFailPos) { return; }
+
+      if (peg$currPos > peg$maxFailPos) {
+        peg$maxFailPos = peg$currPos;
+        peg$maxFailExpected = [];
+      }
+
+      peg$maxFailExpected.push(expected);
+    }
+
+    function peg$buildSimpleError(message, location) {
+      return new peg$SyntaxError(message, null, null, location);
+    }
+
+    function peg$buildStructuredError(expected, found, location) {
+      return new peg$SyntaxError(
+        peg$SyntaxError.buildMessage(expected, found),
+        expected,
+        found,
+        location
+      );
+    }
+
+    function peg$parseFile() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      s1 = peg$parse_();
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parseDCRGraph();
+        if (s2 !== peg$FAILED) {
+          s3 = peg$parse_();
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c0(s2);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+
+      return s0;
+    }
+
+    function peg$parseDCRGraph() {
+      var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9;
+
+      s0 = peg$currPos;
+      peg$savedPos = peg$currPos;
+      s1 = peg$c1();
+      if (s1) {
+        s1 = peg$FAILED;
+      } else {
+        s1 = void 0;
+      }
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parseRelation();
+        if (s2 === peg$FAILED) {
+          s2 = peg$parseEvent();
+          if (s2 === peg$FAILED) {
+            s2 = peg$parseEmpty();
+          }
+        }
+        if (s2 !== peg$FAILED) {
+          s3 = [];
+          s4 = peg$currPos;
+          s5 = peg$parse_();
+          if (s5 !== peg$FAILED) {
+            s6 = peg$parseNxt();
+            if (s6 !== peg$FAILED) {
+              s7 = peg$parse_();
+              if (s7 !== peg$FAILED) {
+                s8 = peg$parseRelation();
+                if (s8 === peg$FAILED) {
+                  s8 = peg$parseEvent();
                 }
-                descriptions.length = j;
-            }
-            switch (descriptions.length) {
-                case 1:
-                    return descriptions[0];
-                case 2:
-                    return descriptions[0] + " or " + descriptions[1];
-                default:
-                    return descriptions.slice(0, -1).join(", ")
-                        + ", or "
-                        + descriptions[descriptions.length - 1];
-            }
-        }
-        function describeFound(found) {
-            return found ? "\"" + literalEscape(found) + "\"" : "end of input";
-        }
-        return "Expected " + describeExpected(expected) + " but " + describeFound(found) + " found.";
-    };
-    function peg$parse(input, options) {
-        options = options !== void 0 ? options : {};
-        var peg$FAILED = {}, peg$startRuleFunctions = { File: peg$parseFile }, peg$startRuleFunction = peg$parseFile, peg$c0 = function (g) { return g; }, peg$c1 = function () { graphs.push(new dcr_1.DCRGraph(top())); }, peg$c2 = function () { return graphs.pop(); }, peg$c3 = "{", peg$c4 = peg$literalExpectation("{", false), peg$c5 = "}", peg$c6 = peg$literalExpectation("}", false), peg$c7 = function (n, l, m, g) { top().addEvent(n, l, m, g); }, peg$c8 = function (n, m, g) { top().addEvent(n, n, m, g); }, peg$c9 = function (n, l, m) { top().addEvent(n, l, m); }, peg$c10 = function (n, m) { top().addEvent(n, n, m); }, peg$c11 = "()", peg$c12 = peg$literalExpectation("()", false), peg$c13 = function () { return { ex: false, in: true, pe: false }; }, peg$c14 = "(", peg$c15 = peg$literalExpectation("(", false), peg$c16 = ")", peg$c17 = peg$literalExpectation(")", false), peg$c18 = function (b1, b2, b3) { return { ex: b1, in: b2, pe: b3 }; }, peg$c19 = "-", peg$c20 = peg$literalExpectation("-", false), peg$c21 = "->*", peg$c22 = peg$literalExpectation("->*", false), peg$c23 = function (es1, d, g, es2) { for (var e1 of es1)
-            for (var e2 of es2)
-                top().addCondition(e1, e2, d, g); }, peg$c24 = "-><>", peg$c25 = peg$literalExpectation("-><>", false), peg$c26 = function (es1, g, es2) { for (var e1 of es1)
-            for (var e2 of es2)
-                top().addMilestone(e1, e2, g); }, peg$c27 = "*-", peg$c28 = peg$literalExpectation("*-", false), peg$c29 = "->", peg$c30 = peg$literalExpectation("->", false), peg$c31 = function (es1, d, g, es2) { for (var e1 of es1)
-            for (var e2 of es2)
-                top().addResponse(e1, e2, d, g); }, peg$c32 = "->+", peg$c33 = peg$literalExpectation("->+", false), peg$c34 = function (es1, g, es2) { for (var e1 of es1)
-            for (var e2 of es2)
-                top().addInclude(e1, e2, g); }, peg$c35 = "->%", peg$c36 = peg$literalExpectation("->%", false), peg$c37 = function (es1, g, es2) { for (var e1 of es1)
-            for (var e2 of es2)
-                top().addExclude(e1, e2, g); }, peg$c38 = function (i) { return i; }, peg$c39 = function () { return undefined; }, peg$c40 = "[", peg$c41 = peg$literalExpectation("[", false), peg$c42 = "]", peg$c43 = peg$literalExpectation("]", false), peg$c44 = function (exp) { return exp; }, peg$c45 = function () { return true; }, peg$c46 = "", peg$c47 = function (n) { return [n]; }, peg$c48 = function (ns) { return ns; }, peg$c49 = function (n, ns) { ns.push(n); return ns; }, peg$c50 = " ", peg$c51 = peg$literalExpectation(" ", false), peg$c52 = function (t, ws, n) { return t.concat(ws.join(''), n); }, peg$c53 = /^[a-z]/, peg$c54 = peg$classExpectation([["a", "z"]], false, false), peg$c55 = /^[A-Z]/, peg$c56 = peg$classExpectation([["A", "Z"]], false, false), peg$c57 = /^[0-9]/, peg$c58 = peg$classExpectation([["0", "9"]], false, false), peg$c59 = "_", peg$c60 = peg$literalExpectation("_", false), peg$c61 = function (str) { return str.join(''); }, peg$c62 = "<", peg$c63 = peg$literalExpectation("<", false), peg$c64 = ">", peg$c65 = peg$literalExpectation(">", false), peg$c66 = function (lab) { return (lab.join('')); }, peg$c67 = "0", peg$c68 = peg$literalExpectation("0", false), peg$c69 = function () { return false; }, peg$c70 = "1", peg$c71 = peg$literalExpectation("1", false), peg$c72 = function () { return true; }, peg$c73 = "false", peg$c74 = peg$literalExpectation("false", false), peg$c75 = "true", peg$c76 = peg$literalExpectation("true", false), peg$c77 = "f", peg$c78 = peg$literalExpectation("f", false), peg$c79 = "t", peg$c80 = peg$literalExpectation("t", false), peg$c81 = peg$otherExpectation("Integer"), peg$c82 = function () { return parseInt(text(), 10); }, peg$c83 = ";", peg$c84 = peg$literalExpectation(";", false), peg$c85 = "\r\n", peg$c86 = peg$literalExpectation("\r\n", false), peg$c87 = "\n", peg$c88 = peg$literalExpectation("\n", false), peg$c89 = peg$anyExpectation(), peg$c90 = ",", peg$c91 = peg$literalExpectation(",", false), peg$c92 = "\t", peg$c93 = peg$literalExpectation("\t", false), peg$currPos = 0, peg$savedPos = 0, peg$posDetailsCache = [{ line: 1, column: 1 }], peg$maxFailPos = 0, peg$maxFailExpected = [], peg$silentFails = 0, peg$result;
-        if ("startRule" in options) {
-            if (!(options.startRule in peg$startRuleFunctions)) {
-                throw new Error("Can't start parsing from rule \"" + options.startRule + "\".");
-            }
-            peg$startRuleFunction = peg$startRuleFunctions[options.startRule];
-        }
-        function text() {
-            return input.substring(peg$savedPos, peg$currPos);
-        }
-        function location() {
-            return peg$computeLocation(peg$savedPos, peg$currPos);
-        }
-        function expected(description, location) {
-            location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos);
-            throw peg$buildStructuredError([peg$otherExpectation(description)], input.substring(peg$savedPos, peg$currPos), location);
-        }
-        function error(message, location) {
-            location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos);
-            throw peg$buildSimpleError(message, location);
-        }
-        function peg$literalExpectation(text, ignoreCase) {
-            return { type: "literal", text: text, ignoreCase: ignoreCase };
-        }
-        function peg$classExpectation(parts, inverted, ignoreCase) {
-            return { type: "class", parts: parts, inverted: inverted, ignoreCase: ignoreCase };
-        }
-        function peg$anyExpectation() {
-            return { type: "any" };
-        }
-        function peg$endExpectation() {
-            return { type: "end" };
-        }
-        function peg$otherExpectation(description) {
-            return { type: "other", description: description };
-        }
-        function peg$computePosDetails(pos) {
-            var details = peg$posDetailsCache[pos], p;
-            if (details) {
-                return details;
-            }
-            else {
-                p = pos - 1;
-                while (!peg$posDetailsCache[p]) {
-                    p--;
+                if (s8 !== peg$FAILED) {
+                  s9 = peg$parse_();
+                  if (s9 !== peg$FAILED) {
+                    s5 = [s5, s6, s7, s8, s9];
+                    s4 = s5;
+                  } else {
+                    peg$currPos = s4;
+                    s4 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s4;
+                  s4 = peg$FAILED;
                 }
-                details = peg$posDetailsCache[p];
-                details = {
-                    line: details.line,
-                    column: details.column
-                };
-                while (p < pos) {
-                    if (input.charCodeAt(p) === 10) {
-                        details.line++;
-                        details.column = 1;
-                    }
-                    else {
-                        details.column++;
-                    }
-                    p++;
+              } else {
+                peg$currPos = s4;
+                s4 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s4;
+              s4 = peg$FAILED;
+            }
+          } else {
+            peg$currPos = s4;
+            s4 = peg$FAILED;
+          }
+          if (s4 === peg$FAILED) {
+            s4 = peg$currPos;
+            s5 = peg$parse_();
+            if (s5 !== peg$FAILED) {
+              s6 = peg$parseEOL();
+              if (s6 !== peg$FAILED) {
+                s7 = peg$parse_();
+                if (s7 !== peg$FAILED) {
+                  s5 = [s5, s6, s7];
+                  s4 = s5;
+                } else {
+                  peg$currPos = s4;
+                  s4 = peg$FAILED;
                 }
-                peg$posDetailsCache[pos] = details;
-                return details;
+              } else {
+                peg$currPos = s4;
+                s4 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s4;
+              s4 = peg$FAILED;
             }
-        }
-        function peg$computeLocation(startPos, endPos) {
-            var startPosDetails = peg$computePosDetails(startPos), endPosDetails = peg$computePosDetails(endPos);
-            return {
-                start: {
-                    offset: startPos,
-                    line: startPosDetails.line,
-                    column: startPosDetails.column
-                },
-                end: {
-                    offset: endPos,
-                    line: endPosDetails.line,
-                    column: endPosDetails.column
-                }
-            };
-        }
-        function peg$fail(expected) {
-            if (peg$currPos < peg$maxFailPos) {
-                return;
-            }
-            if (peg$currPos > peg$maxFailPos) {
-                peg$maxFailPos = peg$currPos;
-                peg$maxFailExpected = [];
-            }
-            peg$maxFailExpected.push(expected);
-        }
-        function peg$buildSimpleError(message, location) {
-            return new peg$SyntaxError(message, null, null, location);
-        }
-        function peg$buildStructuredError(expected, found, location) {
-            return new peg$SyntaxError(peg$SyntaxError.buildMessage(expected, found), expected, found, location);
-        }
-        function peg$parseFile() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            s1 = peg$parse_();
-            if (s1 !== peg$FAILED) {
-                s2 = peg$parseDCRGraph();
-                if (s2 !== peg$FAILED) {
-                    s3 = peg$parse_();
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c0(s2);
-                        s0 = s1;
+          }
+          while (s4 !== peg$FAILED) {
+            s3.push(s4);
+            s4 = peg$currPos;
+            s5 = peg$parse_();
+            if (s5 !== peg$FAILED) {
+              s6 = peg$parseNxt();
+              if (s6 !== peg$FAILED) {
+                s7 = peg$parse_();
+                if (s7 !== peg$FAILED) {
+                  s8 = peg$parseRelation();
+                  if (s8 === peg$FAILED) {
+                    s8 = peg$parseEvent();
+                  }
+                  if (s8 !== peg$FAILED) {
+                    s9 = peg$parse_();
+                    if (s9 !== peg$FAILED) {
+                      s5 = [s5, s6, s7, s8, s9];
+                      s4 = s5;
+                    } else {
+                      peg$currPos = s4;
+                      s4 = peg$FAILED;
                     }
-                    else {
+                  } else {
+                    peg$currPos = s4;
+                    s4 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s4;
+                  s4 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s4;
+                s4 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s4;
+              s4 = peg$FAILED;
+            }
+            if (s4 === peg$FAILED) {
+              s4 = peg$currPos;
+              s5 = peg$parse_();
+              if (s5 !== peg$FAILED) {
+                s6 = peg$parseEOL();
+                if (s6 !== peg$FAILED) {
+                  s7 = peg$parse_();
+                  if (s7 !== peg$FAILED) {
+                    s5 = [s5, s6, s7];
+                    s4 = s5;
+                  } else {
+                    peg$currPos = s4;
+                    s4 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s4;
+                  s4 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s4;
+                s4 = peg$FAILED;
+              }
+            }
+          }
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c2();
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+
+      return s0;
+    }
+
+    function peg$parseEvent() {
+      var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11;
+
+      s0 = peg$currPos;
+      s1 = peg$parseName();
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parse_();
+        if (s2 !== peg$FAILED) {
+          s3 = peg$parseLabel();
+          if (s3 !== peg$FAILED) {
+            s4 = peg$parse_();
+            if (s4 !== peg$FAILED) {
+              s5 = peg$parseMarking();
+              if (s5 !== peg$FAILED) {
+                s6 = peg$parse__();
+                if (s6 !== peg$FAILED) {
+                  if (input.charCodeAt(peg$currPos) === 123) {
+                    s7 = peg$c3;
+                    peg$currPos++;
+                  } else {
+                    s7 = peg$FAILED;
+                    if (peg$silentFails === 0) { peg$fail(peg$c4); }
+                  }
+                  if (s7 !== peg$FAILED) {
+                    s8 = peg$parse__();
+                    if (s8 !== peg$FAILED) {
+                      s9 = peg$parseDCRGraph();
+                      if (s9 !== peg$FAILED) {
+                        s10 = peg$parse__();
+                        if (s10 !== peg$FAILED) {
+                          if (input.charCodeAt(peg$currPos) === 125) {
+                            s11 = peg$c5;
+                            peg$currPos++;
+                          } else {
+                            s11 = peg$FAILED;
+                            if (peg$silentFails === 0) { peg$fail(peg$c6); }
+                          }
+                          if (s11 !== peg$FAILED) {
+                            peg$savedPos = s0;
+                            s1 = peg$c7(s1, s3, s5, s9);
+                            s0 = s1;
+                          } else {
+                            peg$currPos = s0;
+                            s0 = peg$FAILED;
+                          }
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
+                        }
+                      } else {
                         peg$currPos = s0;
                         s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
                     }
-                }
-                else {
+                  } else {
                     peg$currPos = s0;
                     s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-            }
-            else {
+              } else {
                 peg$currPos = s0;
                 s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
             }
-            return s0;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
         }
-        function peg$parseDCRGraph() {
-            var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9;
-            s0 = peg$currPos;
-            peg$savedPos = peg$currPos;
-            s1 = peg$c1();
-            if (s1) {
-                s1 = peg$FAILED;
-            }
-            else {
-                s1 = void 0;
-            }
-            if (s1 !== peg$FAILED) {
-                s2 = peg$parseRelation();
-                if (s2 === peg$FAILED) {
-                    s2 = peg$parseEvent();
-                    if (s2 === peg$FAILED) {
-                        s2 = peg$parseEmpty();
-                    }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        s1 = peg$parseName();
+        if (s1 !== peg$FAILED) {
+          s2 = peg$parse_();
+          if (s2 !== peg$FAILED) {
+            s3 = peg$parseMarking();
+            if (s3 !== peg$FAILED) {
+              s4 = peg$parse__();
+              if (s4 !== peg$FAILED) {
+                if (input.charCodeAt(peg$currPos) === 123) {
+                  s5 = peg$c3;
+                  peg$currPos++;
+                } else {
+                  s5 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c4); }
                 }
-                if (s2 !== peg$FAILED) {
-                    s3 = [];
-                    s4 = peg$currPos;
-                    s5 = peg$parse_();
-                    if (s5 !== peg$FAILED) {
-                        s6 = peg$parseNxt();
-                        if (s6 !== peg$FAILED) {
-                            s7 = peg$parse_();
-                            if (s7 !== peg$FAILED) {
-                                s8 = peg$parseRelation();
-                                if (s8 === peg$FAILED) {
-                                    s8 = peg$parseEvent();
-                                }
-                                if (s8 !== peg$FAILED) {
-                                    s9 = peg$parse_();
-                                    if (s9 !== peg$FAILED) {
-                                        s5 = [s5, s6, s7, s8, s9];
-                                        s4 = s5;
-                                    }
-                                    else {
-                                        peg$currPos = s4;
-                                        s4 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s4;
-                                    s4 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s4;
-                                s4 = peg$FAILED;
-                            }
+                if (s5 !== peg$FAILED) {
+                  s6 = peg$parse__();
+                  if (s6 !== peg$FAILED) {
+                    s7 = peg$parseDCRGraph();
+                    if (s7 !== peg$FAILED) {
+                      s8 = peg$parse__();
+                      if (s8 !== peg$FAILED) {
+                        if (input.charCodeAt(peg$currPos) === 125) {
+                          s9 = peg$c5;
+                          peg$currPos++;
+                        } else {
+                          s9 = peg$FAILED;
+                          if (peg$silentFails === 0) { peg$fail(peg$c6); }
                         }
-                        else {
-                            peg$currPos = s4;
-                            s4 = peg$FAILED;
+                        if (s9 !== peg$FAILED) {
+                          peg$savedPos = s0;
+                          s1 = peg$c8(s1, s3, s7);
+                          s0 = s1;
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
                         }
-                    }
-                    else {
-                        peg$currPos = s4;
-                        s4 = peg$FAILED;
-                    }
-                    if (s4 === peg$FAILED) {
-                        s4 = peg$currPos;
-                        s5 = peg$parse_();
-                        if (s5 !== peg$FAILED) {
-                            s6 = peg$parseEOL();
-                            if (s6 !== peg$FAILED) {
-                                s7 = peg$parse_();
-                                if (s7 !== peg$FAILED) {
-                                    s5 = [s5, s6, s7];
-                                    s4 = s5;
-                                }
-                                else {
-                                    peg$currPos = s4;
-                                    s4 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s4;
-                                s4 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s4;
-                            s4 = peg$FAILED;
-                        }
-                    }
-                    while (s4 !== peg$FAILED) {
-                        s3.push(s4);
-                        s4 = peg$currPos;
-                        s5 = peg$parse_();
-                        if (s5 !== peg$FAILED) {
-                            s6 = peg$parseNxt();
-                            if (s6 !== peg$FAILED) {
-                                s7 = peg$parse_();
-                                if (s7 !== peg$FAILED) {
-                                    s8 = peg$parseRelation();
-                                    if (s8 === peg$FAILED) {
-                                        s8 = peg$parseEvent();
-                                    }
-                                    if (s8 !== peg$FAILED) {
-                                        s9 = peg$parse_();
-                                        if (s9 !== peg$FAILED) {
-                                            s5 = [s5, s6, s7, s8, s9];
-                                            s4 = s5;
-                                        }
-                                        else {
-                                            peg$currPos = s4;
-                                            s4 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s4;
-                                        s4 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s4;
-                                    s4 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s4;
-                                s4 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s4;
-                            s4 = peg$FAILED;
-                        }
-                        if (s4 === peg$FAILED) {
-                            s4 = peg$currPos;
-                            s5 = peg$parse_();
-                            if (s5 !== peg$FAILED) {
-                                s6 = peg$parseEOL();
-                                if (s6 !== peg$FAILED) {
-                                    s7 = peg$parse_();
-                                    if (s7 !== peg$FAILED) {
-                                        s5 = [s5, s6, s7];
-                                        s4 = s5;
-                                    }
-                                    else {
-                                        peg$currPos = s4;
-                                        s4 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s4;
-                                    s4 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s4;
-                                s4 = peg$FAILED;
-                            }
-                        }
-                    }
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c2();
-                        s0 = s1;
-                    }
-                    else {
+                      } else {
                         peg$currPos = s0;
                         s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
                     }
-                }
-                else {
+                  } else {
                     peg$currPos = s0;
                     s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-            }
-            else {
+              } else {
                 peg$currPos = s0;
                 s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
             }
-            return s0;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
         }
-        function peg$parseEvent() {
-            var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11;
+        if (s0 === peg$FAILED) {
+          s0 = peg$currPos;
+          s1 = peg$parseName();
+          if (s1 !== peg$FAILED) {
+            s2 = peg$parse_();
+            if (s2 !== peg$FAILED) {
+              s3 = peg$parseLabel();
+              if (s3 !== peg$FAILED) {
+                s4 = peg$parse_();
+                if (s4 !== peg$FAILED) {
+                  s5 = peg$parseMarking();
+                  if (s5 !== peg$FAILED) {
+                    peg$savedPos = s0;
+                    s1 = peg$c9(s1, s3, s5);
+                    s0 = s1;
+                  } else {
+                    peg$currPos = s0;
+                    s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
+            }
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+          if (s0 === peg$FAILED) {
             s0 = peg$currPos;
             s1 = peg$parseName();
             if (s1 !== peg$FAILED) {
-                s2 = peg$parse_();
-                if (s2 !== peg$FAILED) {
-                    s3 = peg$parseLabel();
-                    if (s3 !== peg$FAILED) {
-                        s4 = peg$parse_();
-                        if (s4 !== peg$FAILED) {
-                            s5 = peg$parseMarking();
-                            if (s5 !== peg$FAILED) {
-                                s6 = peg$parse__();
-                                if (s6 !== peg$FAILED) {
-                                    if (input.charCodeAt(peg$currPos) === 123) {
-                                        s7 = peg$c3;
-                                        peg$currPos++;
-                                    }
-                                    else {
-                                        s7 = peg$FAILED;
-                                        if (peg$silentFails === 0) {
-                                            peg$fail(peg$c4);
-                                        }
-                                    }
-                                    if (s7 !== peg$FAILED) {
-                                        s8 = peg$parse__();
-                                        if (s8 !== peg$FAILED) {
-                                            s9 = peg$parseDCRGraph();
-                                            if (s9 !== peg$FAILED) {
-                                                s10 = peg$parse__();
-                                                if (s10 !== peg$FAILED) {
-                                                    if (input.charCodeAt(peg$currPos) === 125) {
-                                                        s11 = peg$c5;
-                                                        peg$currPos++;
-                                                    }
-                                                    else {
-                                                        s11 = peg$FAILED;
-                                                        if (peg$silentFails === 0) {
-                                                            peg$fail(peg$c6);
-                                                        }
-                                                    }
-                                                    if (s11 !== peg$FAILED) {
-                                                        peg$savedPos = s0;
-                                                        s1 = peg$c7(s1, s3, s5, s9);
-                                                        s0 = s1;
-                                                    }
-                                                    else {
-                                                        peg$currPos = s0;
-                                                        s0 = peg$FAILED;
-                                                    }
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
+              s2 = peg$parse_();
+              if (s2 !== peg$FAILED) {
+                s3 = peg$parseMarking();
+                if (s3 !== peg$FAILED) {
+                  peg$savedPos = s0;
+                  s1 = peg$c10(s1, s3);
+                  s0 = s1;
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
+              } else {
                 peg$currPos = s0;
                 s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
             }
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                s1 = peg$parseName();
-                if (s1 !== peg$FAILED) {
-                    s2 = peg$parse_();
-                    if (s2 !== peg$FAILED) {
-                        s3 = peg$parseMarking();
-                        if (s3 !== peg$FAILED) {
-                            s4 = peg$parse__();
-                            if (s4 !== peg$FAILED) {
-                                if (input.charCodeAt(peg$currPos) === 123) {
-                                    s5 = peg$c3;
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s5 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c4);
-                                    }
-                                }
-                                if (s5 !== peg$FAILED) {
-                                    s6 = peg$parse__();
-                                    if (s6 !== peg$FAILED) {
-                                        s7 = peg$parseDCRGraph();
-                                        if (s7 !== peg$FAILED) {
-                                            s8 = peg$parse__();
-                                            if (s8 !== peg$FAILED) {
-                                                if (input.charCodeAt(peg$currPos) === 125) {
-                                                    s9 = peg$c5;
-                                                    peg$currPos++;
-                                                }
-                                                else {
-                                                    s9 = peg$FAILED;
-                                                    if (peg$silentFails === 0) {
-                                                        peg$fail(peg$c6);
-                                                    }
-                                                }
-                                                if (s9 !== peg$FAILED) {
-                                                    peg$savedPos = s0;
-                                                    s1 = peg$c8(s1, s3, s7);
-                                                    s0 = s1;
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
+          }
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseMarking() {
+      var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9;
+
+      s0 = peg$currPos;
+      if (input.substr(peg$currPos, 2) === peg$c11) {
+        s1 = peg$c11;
+        peg$currPos += 2;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c12); }
+      }
+      if (s1 !== peg$FAILED) {
+        peg$savedPos = s0;
+        s1 = peg$c13();
+      }
+      s0 = s1;
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        if (input.charCodeAt(peg$currPos) === 40) {
+          s1 = peg$c14;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c15); }
+        }
+        if (s1 !== peg$FAILED) {
+          s2 = peg$parse_();
+          if (s2 !== peg$FAILED) {
+            s3 = peg$parseBool();
+            if (s3 !== peg$FAILED) {
+              s4 = peg$parseSep();
+              if (s4 !== peg$FAILED) {
+                s5 = peg$parseBool();
+                if (s5 !== peg$FAILED) {
+                  s6 = peg$parseSep();
+                  if (s6 !== peg$FAILED) {
+                    s7 = peg$parseBool();
+                    if (s7 !== peg$FAILED) {
+                      s8 = peg$parse_();
+                      if (s8 !== peg$FAILED) {
+                        if (input.charCodeAt(peg$currPos) === 41) {
+                          s9 = peg$c16;
+                          peg$currPos++;
+                        } else {
+                          s9 = peg$FAILED;
+                          if (peg$silentFails === 0) { peg$fail(peg$c17); }
                         }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
+                        if (s9 !== peg$FAILED) {
+                          peg$savedPos = s0;
+                          s1 = peg$c18(s3, s5, s7);
+                          s0 = s1;
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
                         }
-                    }
-                    else {
+                      } else {
                         peg$currPos = s0;
                         s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
                     }
-                }
-                else {
+                  } else {
                     peg$currPos = s0;
                     s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-                if (s0 === peg$FAILED) {
-                    s0 = peg$currPos;
-                    s1 = peg$parseName();
-                    if (s1 !== peg$FAILED) {
-                        s2 = peg$parse_();
-                        if (s2 !== peg$FAILED) {
-                            s3 = peg$parseLabel();
-                            if (s3 !== peg$FAILED) {
-                                s4 = peg$parse_();
-                                if (s4 !== peg$FAILED) {
-                                    s5 = peg$parseMarking();
-                                    if (s5 !== peg$FAILED) {
-                                        peg$savedPos = s0;
-                                        s1 = peg$c9(s1, s3, s5);
-                                        s0 = s1;
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                    if (s0 === peg$FAILED) {
-                        s0 = peg$currPos;
-                        s1 = peg$parseName();
-                        if (s1 !== peg$FAILED) {
-                            s2 = peg$parse_();
-                            if (s2 !== peg$FAILED) {
-                                s3 = peg$parseMarking();
-                                if (s3 !== peg$FAILED) {
-                                    peg$savedPos = s0;
-                                    s1 = peg$c10(s1, s3);
-                                    s0 = s1;
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
             }
-            return s0;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
         }
-        function peg$parseMarking() {
-            var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9;
-            s0 = peg$currPos;
-            if (input.substr(peg$currPos, 2) === peg$c11) {
-                s1 = peg$c11;
+        if (s0 === peg$FAILED) {
+          s0 = peg$currPos;
+          s1 = peg$parseEmpty();
+          if (s1 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c13();
+          }
+          s0 = s1;
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseRelation() {
+      var s0, s1, s2, s3, s4, s5, s6, s7, s8;
+
+      s0 = peg$currPos;
+      s1 = peg$parseEventids();
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parse_();
+        if (s2 !== peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 45) {
+            s3 = peg$c19;
+            peg$currPos++;
+          } else {
+            s3 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c20); }
+          }
+          if (s3 !== peg$FAILED) {
+            s4 = peg$parseTime();
+            if (s4 !== peg$FAILED) {
+              s5 = peg$parseGuard();
+              if (s5 !== peg$FAILED) {
+                if (input.substr(peg$currPos, 3) === peg$c21) {
+                  s6 = peg$c21;
+                  peg$currPos += 3;
+                } else {
+                  s6 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c22); }
+                }
+                if (s6 !== peg$FAILED) {
+                  s7 = peg$parse_();
+                  if (s7 !== peg$FAILED) {
+                    s8 = peg$parseEventids();
+                    if (s8 !== peg$FAILED) {
+                      peg$savedPos = s0;
+                      s1 = peg$c23(s1, s4, s5, s8);
+                      s0 = s1;
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
+                    }
+                  } else {
+                    peg$currPos = s0;
+                    s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
+            }
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        s1 = peg$parseEventids();
+        if (s1 !== peg$FAILED) {
+          s2 = peg$parse_();
+          if (s2 !== peg$FAILED) {
+            if (input.charCodeAt(peg$currPos) === 45) {
+              s3 = peg$c19;
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c20); }
+            }
+            if (s3 !== peg$FAILED) {
+              s4 = peg$parseGuard();
+              if (s4 !== peg$FAILED) {
+                if (input.substr(peg$currPos, 4) === peg$c24) {
+                  s5 = peg$c24;
+                  peg$currPos += 4;
+                } else {
+                  s5 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c25); }
+                }
+                if (s5 !== peg$FAILED) {
+                  s6 = peg$parse_();
+                  if (s6 !== peg$FAILED) {
+                    s7 = peg$parseEventids();
+                    if (s7 !== peg$FAILED) {
+                      peg$savedPos = s0;
+                      s1 = peg$c26(s1, s4, s7);
+                      s0 = s1;
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
+                    }
+                  } else {
+                    peg$currPos = s0;
+                    s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
+            }
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+        if (s0 === peg$FAILED) {
+          s0 = peg$currPos;
+          s1 = peg$parseEventids();
+          if (s1 !== peg$FAILED) {
+            s2 = peg$parse_();
+            if (s2 !== peg$FAILED) {
+              if (input.substr(peg$currPos, 2) === peg$c27) {
+                s3 = peg$c27;
                 peg$currPos += 2;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c12);
-                }
-            }
-            if (s1 !== peg$FAILED) {
-                peg$savedPos = s0;
-                s1 = peg$c13();
-            }
-            s0 = s1;
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                if (input.charCodeAt(peg$currPos) === 40) {
-                    s1 = peg$c14;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c15);
+              } else {
+                s3 = peg$FAILED;
+                if (peg$silentFails === 0) { peg$fail(peg$c28); }
+              }
+              if (s3 !== peg$FAILED) {
+                s4 = peg$parseTime();
+                if (s4 !== peg$FAILED) {
+                  s5 = peg$parseGuard();
+                  if (s5 !== peg$FAILED) {
+                    if (input.substr(peg$currPos, 2) === peg$c29) {
+                      s6 = peg$c29;
+                      peg$currPos += 2;
+                    } else {
+                      s6 = peg$FAILED;
+                      if (peg$silentFails === 0) { peg$fail(peg$c30); }
                     }
-                }
-                if (s1 !== peg$FAILED) {
-                    s2 = peg$parse_();
-                    if (s2 !== peg$FAILED) {
-                        s3 = peg$parseBool();
-                        if (s3 !== peg$FAILED) {
-                            s4 = peg$parseSep();
-                            if (s4 !== peg$FAILED) {
-                                s5 = peg$parseBool();
-                                if (s5 !== peg$FAILED) {
-                                    s6 = peg$parseSep();
-                                    if (s6 !== peg$FAILED) {
-                                        s7 = peg$parseBool();
-                                        if (s7 !== peg$FAILED) {
-                                            s8 = peg$parse_();
-                                            if (s8 !== peg$FAILED) {
-                                                if (input.charCodeAt(peg$currPos) === 41) {
-                                                    s9 = peg$c16;
-                                                    peg$currPos++;
-                                                }
-                                                else {
-                                                    s9 = peg$FAILED;
-                                                    if (peg$silentFails === 0) {
-                                                        peg$fail(peg$c17);
-                                                    }
-                                                }
-                                                if (s9 !== peg$FAILED) {
-                                                    peg$savedPos = s0;
-                                                    s1 = peg$c18(s3, s5, s7);
-                                                    s0 = s1;
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
+                    if (s6 !== peg$FAILED) {
+                      s7 = peg$parse_();
+                      if (s7 !== peg$FAILED) {
+                        s8 = peg$parseEventids();
+                        if (s8 !== peg$FAILED) {
+                          peg$savedPos = s0;
+                          s1 = peg$c31(s1, s4, s5, s8);
+                          s0 = s1;
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
                         }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
+                      } else {
                         peg$currPos = s0;
                         s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
                     }
-                }
-                else {
+                  } else {
                     peg$currPos = s0;
                     s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-                if (s0 === peg$FAILED) {
-                    s0 = peg$currPos;
-                    s1 = peg$parseEmpty();
-                    if (s1 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c13();
-                    }
-                    s0 = s1;
-                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
             }
-            return s0;
-        }
-        function peg$parseRelation() {
-            var s0, s1, s2, s3, s4, s5, s6, s7, s8;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+          if (s0 === peg$FAILED) {
             s0 = peg$currPos;
             s1 = peg$parseEventids();
             if (s1 !== peg$FAILED) {
+              s2 = peg$parse_();
+              if (s2 !== peg$FAILED) {
+                if (input.charCodeAt(peg$currPos) === 45) {
+                  s3 = peg$c19;
+                  peg$currPos++;
+                } else {
+                  s3 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c20); }
+                }
+                if (s3 !== peg$FAILED) {
+                  s4 = peg$parseGuard();
+                  if (s4 !== peg$FAILED) {
+                    if (input.substr(peg$currPos, 3) === peg$c32) {
+                      s5 = peg$c32;
+                      peg$currPos += 3;
+                    } else {
+                      s5 = peg$FAILED;
+                      if (peg$silentFails === 0) { peg$fail(peg$c33); }
+                    }
+                    if (s5 !== peg$FAILED) {
+                      s6 = peg$parse_();
+                      if (s6 !== peg$FAILED) {
+                        s7 = peg$parseEventids();
+                        if (s7 !== peg$FAILED) {
+                          peg$savedPos = s0;
+                          s1 = peg$c34(s1, s4, s7);
+                          s0 = s1;
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
+                        }
+                      } else {
+                        peg$currPos = s0;
+                        s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
+                    }
+                  } else {
+                    peg$currPos = s0;
+                    s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
+                }
+              } else {
+                peg$currPos = s0;
+                s0 = peg$FAILED;
+              }
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
+            }
+            if (s0 === peg$FAILED) {
+              s0 = peg$currPos;
+              s1 = peg$parseEventids();
+              if (s1 !== peg$FAILED) {
                 s2 = peg$parse_();
                 if (s2 !== peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 45) {
-                        s3 = peg$c19;
-                        peg$currPos++;
-                    }
-                    else {
-                        s3 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c20);
-                        }
-                    }
-                    if (s3 !== peg$FAILED) {
-                        s4 = peg$parseTime();
-                        if (s4 !== peg$FAILED) {
-                            s5 = peg$parseGuard();
-                            if (s5 !== peg$FAILED) {
-                                if (input.substr(peg$currPos, 3) === peg$c21) {
-                                    s6 = peg$c21;
-                                    peg$currPos += 3;
-                                }
-                                else {
-                                    s6 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c22);
-                                    }
-                                }
-                                if (s6 !== peg$FAILED) {
-                                    s7 = peg$parse_();
-                                    if (s7 !== peg$FAILED) {
-                                        s8 = peg$parseEventids();
-                                        if (s8 !== peg$FAILED) {
-                                            peg$savedPos = s0;
-                                            s1 = peg$c23(s1, s4, s5, s8);
-                                            s0 = s1;
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                s1 = peg$parseEventids();
-                if (s1 !== peg$FAILED) {
-                    s2 = peg$parse_();
-                    if (s2 !== peg$FAILED) {
-                        if (input.charCodeAt(peg$currPos) === 45) {
-                            s3 = peg$c19;
-                            peg$currPos++;
-                        }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c20);
-                            }
-                        }
-                        if (s3 !== peg$FAILED) {
-                            s4 = peg$parseGuard();
-                            if (s4 !== peg$FAILED) {
-                                if (input.substr(peg$currPos, 4) === peg$c24) {
-                                    s5 = peg$c24;
-                                    peg$currPos += 4;
-                                }
-                                else {
-                                    s5 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c25);
-                                    }
-                                }
-                                if (s5 !== peg$FAILED) {
-                                    s6 = peg$parse_();
-                                    if (s6 !== peg$FAILED) {
-                                        s7 = peg$parseEventids();
-                                        if (s7 !== peg$FAILED) {
-                                            peg$savedPos = s0;
-                                            s1 = peg$c26(s1, s4, s7);
-                                            s0 = s1;
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-                if (s0 === peg$FAILED) {
-                    s0 = peg$currPos;
-                    s1 = peg$parseEventids();
-                    if (s1 !== peg$FAILED) {
-                        s2 = peg$parse_();
-                        if (s2 !== peg$FAILED) {
-                            if (input.substr(peg$currPos, 2) === peg$c27) {
-                                s3 = peg$c27;
-                                peg$currPos += 2;
-                            }
-                            else {
-                                s3 = peg$FAILED;
-                                if (peg$silentFails === 0) {
-                                    peg$fail(peg$c28);
-                                }
-                            }
-                            if (s3 !== peg$FAILED) {
-                                s4 = peg$parseTime();
-                                if (s4 !== peg$FAILED) {
-                                    s5 = peg$parseGuard();
-                                    if (s5 !== peg$FAILED) {
-                                        if (input.substr(peg$currPos, 2) === peg$c29) {
-                                            s6 = peg$c29;
-                                            peg$currPos += 2;
-                                        }
-                                        else {
-                                            s6 = peg$FAILED;
-                                            if (peg$silentFails === 0) {
-                                                peg$fail(peg$c30);
-                                            }
-                                        }
-                                        if (s6 !== peg$FAILED) {
-                                            s7 = peg$parse_();
-                                            if (s7 !== peg$FAILED) {
-                                                s8 = peg$parseEventids();
-                                                if (s8 !== peg$FAILED) {
-                                                    peg$savedPos = s0;
-                                                    s1 = peg$c31(s1, s4, s5, s8);
-                                                    s0 = s1;
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                    if (s0 === peg$FAILED) {
-                        s0 = peg$currPos;
-                        s1 = peg$parseEventids();
-                        if (s1 !== peg$FAILED) {
-                            s2 = peg$parse_();
-                            if (s2 !== peg$FAILED) {
-                                if (input.charCodeAt(peg$currPos) === 45) {
-                                    s3 = peg$c19;
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s3 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c20);
-                                    }
-                                }
-                                if (s3 !== peg$FAILED) {
-                                    s4 = peg$parseGuard();
-                                    if (s4 !== peg$FAILED) {
-                                        if (input.substr(peg$currPos, 3) === peg$c32) {
-                                            s5 = peg$c32;
-                                            peg$currPos += 3;
-                                        }
-                                        else {
-                                            s5 = peg$FAILED;
-                                            if (peg$silentFails === 0) {
-                                                peg$fail(peg$c33);
-                                            }
-                                        }
-                                        if (s5 !== peg$FAILED) {
-                                            s6 = peg$parse_();
-                                            if (s6 !== peg$FAILED) {
-                                                s7 = peg$parseEventids();
-                                                if (s7 !== peg$FAILED) {
-                                                    peg$savedPos = s0;
-                                                    s1 = peg$c34(s1, s4, s7);
-                                                    s0 = s1;
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                        if (s0 === peg$FAILED) {
-                            s0 = peg$currPos;
-                            s1 = peg$parseEventids();
-                            if (s1 !== peg$FAILED) {
-                                s2 = peg$parse_();
-                                if (s2 !== peg$FAILED) {
-                                    if (input.charCodeAt(peg$currPos) === 45) {
-                                        s3 = peg$c19;
-                                        peg$currPos++;
-                                    }
-                                    else {
-                                        s3 = peg$FAILED;
-                                        if (peg$silentFails === 0) {
-                                            peg$fail(peg$c20);
-                                        }
-                                    }
-                                    if (s3 !== peg$FAILED) {
-                                        s4 = peg$parseGuard();
-                                        if (s4 !== peg$FAILED) {
-                                            if (input.substr(peg$currPos, 3) === peg$c35) {
-                                                s5 = peg$c35;
-                                                peg$currPos += 3;
-                                            }
-                                            else {
-                                                s5 = peg$FAILED;
-                                                if (peg$silentFails === 0) {
-                                                    peg$fail(peg$c36);
-                                                }
-                                            }
-                                            if (s5 !== peg$FAILED) {
-                                                s6 = peg$parse_();
-                                                if (s6 !== peg$FAILED) {
-                                                    s7 = peg$parseEventids();
-                                                    if (s7 !== peg$FAILED) {
-                                                        peg$savedPos = s0;
-                                                        s1 = peg$c37(s1, s4, s7);
-                                                        s0 = s1;
-                                                    }
-                                                    else {
-                                                        peg$currPos = s0;
-                                                        s0 = peg$FAILED;
-                                                    }
-                                                }
-                                                else {
-                                                    peg$currPos = s0;
-                                                    s0 = peg$FAILED;
-                                                }
-                                            }
-                                            else {
-                                                peg$currPos = s0;
-                                                s0 = peg$FAILED;
-                                            }
-                                        }
-                                        else {
-                                            peg$currPos = s0;
-                                            s0 = peg$FAILED;
-                                        }
-                                    }
-                                    else {
-                                        peg$currPos = s0;
-                                        s0 = peg$FAILED;
-                                    }
-                                }
-                                else {
-                                    peg$currPos = s0;
-                                    s0 = peg$FAILED;
-                                }
-                            }
-                            else {
-                                peg$currPos = s0;
-                                s0 = peg$FAILED;
-                            }
-                        }
-                    }
-                }
-            }
-            return s0;
-        }
-        function peg$parseTime() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            if (input.charCodeAt(peg$currPos) === 40) {
-                s1 = peg$c14;
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c15);
-                }
-            }
-            if (s1 !== peg$FAILED) {
-                s2 = peg$parseInteger();
-                if (s2 !== peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 41) {
-                        s3 = peg$c16;
-                        peg$currPos++;
-                    }
-                    else {
-                        s3 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c17);
-                        }
-                    }
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c38(s2);
-                        s0 = s1;
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                s1 = peg$parseEmpty();
-                if (s1 !== peg$FAILED) {
-                    peg$savedPos = s0;
-                    s1 = peg$c39();
-                }
-                s0 = s1;
-            }
-            return s0;
-        }
-        function peg$parseGuard() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            if (input.charCodeAt(peg$currPos) === 91) {
-                s1 = peg$c40;
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c41);
-                }
-            }
-            if (s1 !== peg$FAILED) {
-                s2 = peg$parseExpression();
-                if (s2 !== peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 93) {
-                        s3 = peg$c42;
-                        peg$currPos++;
-                    }
-                    else {
-                        s3 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c43);
-                        }
-                    }
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c44(s2);
-                        s0 = s1;
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                s1 = peg$parseEmpty();
-                if (s1 !== peg$FAILED) {
-                    peg$savedPos = s0;
-                    s1 = peg$c45();
-                }
-                s0 = s1;
-            }
-            return s0;
-        }
-        function peg$parseExpression() {
-            var s0;
-            s0 = peg$c46;
-            return s0;
-        }
-        function peg$parseEventids() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            s1 = peg$parseName();
-            if (s1 !== peg$FAILED) {
-                peg$savedPos = s0;
-                s1 = peg$c47(s1);
-            }
-            s0 = s1;
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                if (input.charCodeAt(peg$currPos) === 40) {
-                    s1 = peg$c14;
+                  if (input.charCodeAt(peg$currPos) === 45) {
+                    s3 = peg$c19;
                     peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c15);
-                    }
-                }
-                if (s1 !== peg$FAILED) {
-                    s2 = peg$parseNames();
-                    if (s2 !== peg$FAILED) {
-                        if (input.charCodeAt(peg$currPos) === 41) {
-                            s3 = peg$c16;
-                            peg$currPos++;
-                        }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c17);
-                            }
-                        }
-                        if (s3 !== peg$FAILED) {
-                            peg$savedPos = s0;
-                            s1 = peg$c48(s2);
-                            s0 = s1;
-                        }
-                        else {
-                            peg$currPos = s0;
-                            s0 = peg$FAILED;
-                        }
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            return s0;
-        }
-        function peg$parseNames() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            s1 = peg$parseName();
-            if (s1 !== peg$FAILED) {
-                s2 = peg$parseSep();
-                if (s2 !== peg$FAILED) {
-                    s3 = peg$parseNames();
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c49(s1, s3);
-                        s0 = s1;
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            if (s0 === peg$FAILED) {
-                s0 = peg$currPos;
-                s1 = peg$parseName();
-                if (s1 !== peg$FAILED) {
-                    peg$savedPos = s0;
-                    s1 = peg$c47(s1);
-                }
-                s0 = s1;
-            }
-            return s0;
-        }
-        function peg$parseName() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            s1 = peg$parseText();
-            if (s1 !== peg$FAILED) {
-                s2 = [];
-                if (input.charCodeAt(peg$currPos) === 32) {
-                    s3 = peg$c50;
-                    peg$currPos++;
-                }
-                else {
+                  } else {
                     s3 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c51);
-                    }
-                }
-                if (s3 !== peg$FAILED) {
-                    while (s3 !== peg$FAILED) {
-                        s2.push(s3);
-                        if (input.charCodeAt(peg$currPos) === 32) {
-                            s3 = peg$c50;
-                            peg$currPos++;
+                    if (peg$silentFails === 0) { peg$fail(peg$c20); }
+                  }
+                  if (s3 !== peg$FAILED) {
+                    s4 = peg$parseGuard();
+                    if (s4 !== peg$FAILED) {
+                      if (input.substr(peg$currPos, 3) === peg$c35) {
+                        s5 = peg$c35;
+                        peg$currPos += 3;
+                      } else {
+                        s5 = peg$FAILED;
+                        if (peg$silentFails === 0) { peg$fail(peg$c36); }
+                      }
+                      if (s5 !== peg$FAILED) {
+                        s6 = peg$parse_();
+                        if (s6 !== peg$FAILED) {
+                          s7 = peg$parseEventids();
+                          if (s7 !== peg$FAILED) {
+                            peg$savedPos = s0;
+                            s1 = peg$c37(s1, s4, s7);
+                            s0 = s1;
+                          } else {
+                            peg$currPos = s0;
+                            s0 = peg$FAILED;
+                          }
+                        } else {
+                          peg$currPos = s0;
+                          s0 = peg$FAILED;
                         }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c51);
-                            }
-                        }
-                    }
-                }
-                else {
-                    s2 = peg$FAILED;
-                }
-                if (s2 !== peg$FAILED) {
-                    s3 = peg$parseName();
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c52(s1, s2, s3);
-                        s0 = s1;
-                    }
-                    else {
+                      } else {
                         peg$currPos = s0;
                         s0 = peg$FAILED;
+                      }
+                    } else {
+                      peg$currPos = s0;
+                      s0 = peg$FAILED;
                     }
-                }
-                else {
+                  } else {
                     peg$currPos = s0;
                     s0 = peg$FAILED;
+                  }
+                } else {
+                  peg$currPos = s0;
+                  s0 = peg$FAILED;
                 }
-            }
-            else {
+              } else {
                 peg$currPos = s0;
                 s0 = peg$FAILED;
+              }
             }
-            if (s0 === peg$FAILED) {
-                s0 = peg$parseText();
-            }
-            return s0;
+          }
         }
-        function peg$parseText() {
-            var s0, s1, s2;
-            s0 = peg$currPos;
-            s1 = [];
-            if (peg$c53.test(input.charAt(peg$currPos))) {
-                s2 = input.charAt(peg$currPos);
-                peg$currPos++;
+      }
+
+      return s0;
+    }
+
+    function peg$parseTime() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      if (input.charCodeAt(peg$currPos) === 40) {
+        s1 = peg$c14;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c15); }
+      }
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parseInteger();
+        if (s2 !== peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 41) {
+            s3 = peg$c16;
+            peg$currPos++;
+          } else {
+            s3 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c17); }
+          }
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c38(s2);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        s1 = peg$parseEmpty();
+        if (s1 !== peg$FAILED) {
+          peg$savedPos = s0;
+          s1 = peg$c39();
+        }
+        s0 = s1;
+      }
+
+      return s0;
+    }
+
+    function peg$parseGuard() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      if (input.charCodeAt(peg$currPos) === 91) {
+        s1 = peg$c40;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c41); }
+      }
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parseExpression();
+        if (s2 !== peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 93) {
+            s3 = peg$c42;
+            peg$currPos++;
+          } else {
+            s3 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c43); }
+          }
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c44(s2);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        s1 = peg$parseEmpty();
+        if (s1 !== peg$FAILED) {
+          peg$savedPos = s0;
+          s1 = peg$c45();
+        }
+        s0 = s1;
+      }
+
+      return s0;
+    }
+
+    function peg$parseExpression() {
+      var s0;
+
+      s0 = peg$c46;
+
+      return s0;
+    }
+
+    function peg$parseEventids() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      s1 = peg$parseName();
+      if (s1 !== peg$FAILED) {
+        peg$savedPos = s0;
+        s1 = peg$c47(s1);
+      }
+      s0 = s1;
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        if (input.charCodeAt(peg$currPos) === 40) {
+          s1 = peg$c14;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c15); }
+        }
+        if (s1 !== peg$FAILED) {
+          s2 = peg$parseNames();
+          if (s2 !== peg$FAILED) {
+            if (input.charCodeAt(peg$currPos) === 41) {
+              s3 = peg$c16;
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c17); }
             }
-            else {
-                s2 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c54);
-                }
+            if (s3 !== peg$FAILED) {
+              peg$savedPos = s0;
+              s1 = peg$c48(s2);
+              s0 = s1;
+            } else {
+              peg$currPos = s0;
+              s0 = peg$FAILED;
+            }
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseNames() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      s1 = peg$parseName();
+      if (s1 !== peg$FAILED) {
+        s2 = peg$parseSep();
+        if (s2 !== peg$FAILED) {
+          s3 = peg$parseNames();
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c49(s1, s3);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        s1 = peg$parseName();
+        if (s1 !== peg$FAILED) {
+          peg$savedPos = s0;
+          s1 = peg$c47(s1);
+        }
+        s0 = s1;
+      }
+
+      return s0;
+    }
+
+    function peg$parseName() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      s1 = peg$parseText();
+      if (s1 !== peg$FAILED) {
+        s2 = [];
+        if (input.charCodeAt(peg$currPos) === 32) {
+          s3 = peg$c50;
+          peg$currPos++;
+        } else {
+          s3 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c51); }
+        }
+        if (s3 !== peg$FAILED) {
+          while (s3 !== peg$FAILED) {
+            s2.push(s3);
+            if (input.charCodeAt(peg$currPos) === 32) {
+              s3 = peg$c50;
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c51); }
+            }
+          }
+        } else {
+          s2 = peg$FAILED;
+        }
+        if (s2 !== peg$FAILED) {
+          s3 = peg$parseName();
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c52(s1, s2, s3);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      if (s0 === peg$FAILED) {
+        s0 = peg$parseText();
+      }
+
+      return s0;
+    }
+
+    function peg$parseText() {
+      var s0, s1, s2;
+
+      s0 = peg$currPos;
+      s1 = [];
+      if (peg$c53.test(input.charAt(peg$currPos))) {
+        s2 = input.charAt(peg$currPos);
+        peg$currPos++;
+      } else {
+        s2 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c54); }
+      }
+      if (s2 === peg$FAILED) {
+        if (peg$c55.test(input.charAt(peg$currPos))) {
+          s2 = input.charAt(peg$currPos);
+          peg$currPos++;
+        } else {
+          s2 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c56); }
+        }
+        if (s2 === peg$FAILED) {
+          if (peg$c57.test(input.charAt(peg$currPos))) {
+            s2 = input.charAt(peg$currPos);
+            peg$currPos++;
+          } else {
+            s2 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c58); }
+          }
+          if (s2 === peg$FAILED) {
+            if (input.charCodeAt(peg$currPos) === 95) {
+              s2 = peg$c59;
+              peg$currPos++;
+            } else {
+              s2 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c60); }
+            }
+          }
+        }
+      }
+      if (s2 !== peg$FAILED) {
+        while (s2 !== peg$FAILED) {
+          s1.push(s2);
+          if (peg$c53.test(input.charAt(peg$currPos))) {
+            s2 = input.charAt(peg$currPos);
+            peg$currPos++;
+          } else {
+            s2 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c54); }
+          }
+          if (s2 === peg$FAILED) {
+            if (peg$c55.test(input.charAt(peg$currPos))) {
+              s2 = input.charAt(peg$currPos);
+              peg$currPos++;
+            } else {
+              s2 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c56); }
             }
             if (s2 === peg$FAILED) {
-                if (peg$c55.test(input.charAt(peg$currPos))) {
-                    s2 = input.charAt(peg$currPos);
-                    peg$currPos++;
-                }
-                else {
-                    s2 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c56);
-                    }
-                }
-                if (s2 === peg$FAILED) {
-                    if (peg$c57.test(input.charAt(peg$currPos))) {
-                        s2 = input.charAt(peg$currPos);
-                        peg$currPos++;
-                    }
-                    else {
-                        s2 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c58);
-                        }
-                    }
-                    if (s2 === peg$FAILED) {
-                        if (input.charCodeAt(peg$currPos) === 95) {
-                            s2 = peg$c59;
-                            peg$currPos++;
-                        }
-                        else {
-                            s2 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c60);
-                            }
-                        }
-                    }
-                }
-            }
-            if (s2 !== peg$FAILED) {
-                while (s2 !== peg$FAILED) {
-                    s1.push(s2);
-                    if (peg$c53.test(input.charAt(peg$currPos))) {
-                        s2 = input.charAt(peg$currPos);
-                        peg$currPos++;
-                    }
-                    else {
-                        s2 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c54);
-                        }
-                    }
-                    if (s2 === peg$FAILED) {
-                        if (peg$c55.test(input.charAt(peg$currPos))) {
-                            s2 = input.charAt(peg$currPos);
-                            peg$currPos++;
-                        }
-                        else {
-                            s2 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c56);
-                            }
-                        }
-                        if (s2 === peg$FAILED) {
-                            if (peg$c57.test(input.charAt(peg$currPos))) {
-                                s2 = input.charAt(peg$currPos);
-                                peg$currPos++;
-                            }
-                            else {
-                                s2 = peg$FAILED;
-                                if (peg$silentFails === 0) {
-                                    peg$fail(peg$c58);
-                                }
-                            }
-                            if (s2 === peg$FAILED) {
-                                if (input.charCodeAt(peg$currPos) === 95) {
-                                    s2 = peg$c59;
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s2 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c60);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            else {
-                s1 = peg$FAILED;
-            }
-            if (s1 !== peg$FAILED) {
-                peg$savedPos = s0;
-                s1 = peg$c61(s1);
-            }
-            s0 = s1;
-            return s0;
-        }
-        function peg$parseLabel() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            if (input.charCodeAt(peg$currPos) === 60) {
-                s1 = peg$c62;
+              if (peg$c57.test(input.charAt(peg$currPos))) {
+                s2 = input.charAt(peg$currPos);
                 peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c63);
+              } else {
+                s2 = peg$FAILED;
+                if (peg$silentFails === 0) { peg$fail(peg$c58); }
+              }
+              if (s2 === peg$FAILED) {
+                if (input.charCodeAt(peg$currPos) === 95) {
+                  s2 = peg$c59;
+                  peg$currPos++;
+                } else {
+                  s2 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c60); }
                 }
+              }
             }
-            if (s1 !== peg$FAILED) {
-                s2 = [];
-                if (peg$c53.test(input.charAt(peg$currPos))) {
-                    s3 = input.charAt(peg$currPos);
-                    peg$currPos++;
-                }
-                else {
-                    s3 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c54);
-                    }
+          }
+        }
+      } else {
+        s1 = peg$FAILED;
+      }
+      if (s1 !== peg$FAILED) {
+        peg$savedPos = s0;
+        s1 = peg$c61(s1);
+      }
+      s0 = s1;
+
+      return s0;
+    }
+
+    function peg$parseLabel() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      if (input.charCodeAt(peg$currPos) === 60) {
+        s1 = peg$c62;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c63); }
+      }
+      if (s1 !== peg$FAILED) {
+        s2 = [];
+        if (peg$c53.test(input.charAt(peg$currPos))) {
+          s3 = input.charAt(peg$currPos);
+          peg$currPos++;
+        } else {
+          s3 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c54); }
+        }
+        if (s3 === peg$FAILED) {
+          if (peg$c55.test(input.charAt(peg$currPos))) {
+            s3 = input.charAt(peg$currPos);
+            peg$currPos++;
+          } else {
+            s3 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c56); }
+          }
+          if (s3 === peg$FAILED) {
+            if (peg$c57.test(input.charAt(peg$currPos))) {
+              s3 = input.charAt(peg$currPos);
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c58); }
+            }
+            if (s3 === peg$FAILED) {
+              if (input.charCodeAt(peg$currPos) === 95) {
+                s3 = peg$c59;
+                peg$currPos++;
+              } else {
+                s3 = peg$FAILED;
+                if (peg$silentFails === 0) { peg$fail(peg$c60); }
+              }
+              if (s3 === peg$FAILED) {
+                if (input.charCodeAt(peg$currPos) === 45) {
+                  s3 = peg$c19;
+                  peg$currPos++;
+                } else {
+                  s3 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c20); }
                 }
                 if (s3 === peg$FAILED) {
-                    if (peg$c55.test(input.charAt(peg$currPos))) {
-                        s3 = input.charAt(peg$currPos);
-                        peg$currPos++;
-                    }
-                    else {
-                        s3 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c56);
-                        }
+                  if (input.charCodeAt(peg$currPos) === 32) {
+                    s3 = peg$c50;
+                    peg$currPos++;
+                  } else {
+                    s3 = peg$FAILED;
+                    if (peg$silentFails === 0) { peg$fail(peg$c51); }
+                  }
+                }
+              }
+            }
+          }
+        }
+        if (s3 !== peg$FAILED) {
+          while (s3 !== peg$FAILED) {
+            s2.push(s3);
+            if (peg$c53.test(input.charAt(peg$currPos))) {
+              s3 = input.charAt(peg$currPos);
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c54); }
+            }
+            if (s3 === peg$FAILED) {
+              if (peg$c55.test(input.charAt(peg$currPos))) {
+                s3 = input.charAt(peg$currPos);
+                peg$currPos++;
+              } else {
+                s3 = peg$FAILED;
+                if (peg$silentFails === 0) { peg$fail(peg$c56); }
+              }
+              if (s3 === peg$FAILED) {
+                if (peg$c57.test(input.charAt(peg$currPos))) {
+                  s3 = input.charAt(peg$currPos);
+                  peg$currPos++;
+                } else {
+                  s3 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c58); }
+                }
+                if (s3 === peg$FAILED) {
+                  if (input.charCodeAt(peg$currPos) === 95) {
+                    s3 = peg$c59;
+                    peg$currPos++;
+                  } else {
+                    s3 = peg$FAILED;
+                    if (peg$silentFails === 0) { peg$fail(peg$c60); }
+                  }
+                  if (s3 === peg$FAILED) {
+                    if (input.charCodeAt(peg$currPos) === 45) {
+                      s3 = peg$c19;
+                      peg$currPos++;
+                    } else {
+                      s3 = peg$FAILED;
+                      if (peg$silentFails === 0) { peg$fail(peg$c20); }
                     }
                     if (s3 === peg$FAILED) {
-                        if (peg$c57.test(input.charAt(peg$currPos))) {
-                            s3 = input.charAt(peg$currPos);
-                            peg$currPos++;
-                        }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c58);
-                            }
-                        }
-                        if (s3 === peg$FAILED) {
-                            if (input.charCodeAt(peg$currPos) === 95) {
-                                s3 = peg$c59;
-                                peg$currPos++;
-                            }
-                            else {
-                                s3 = peg$FAILED;
-                                if (peg$silentFails === 0) {
-                                    peg$fail(peg$c60);
-                                }
-                            }
-                            if (s3 === peg$FAILED) {
-                                if (input.charCodeAt(peg$currPos) === 45) {
-                                    s3 = peg$c19;
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s3 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c20);
-                                    }
-                                }
-                                if (s3 === peg$FAILED) {
-                                    if (input.charCodeAt(peg$currPos) === 32) {
-                                        s3 = peg$c50;
-                                        peg$currPos++;
-                                    }
-                                    else {
-                                        s3 = peg$FAILED;
-                                        if (peg$silentFails === 0) {
-                                            peg$fail(peg$c51);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                if (s3 !== peg$FAILED) {
-                    while (s3 !== peg$FAILED) {
-                        s2.push(s3);
-                        if (peg$c53.test(input.charAt(peg$currPos))) {
-                            s3 = input.charAt(peg$currPos);
-                            peg$currPos++;
-                        }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c54);
-                            }
-                        }
-                        if (s3 === peg$FAILED) {
-                            if (peg$c55.test(input.charAt(peg$currPos))) {
-                                s3 = input.charAt(peg$currPos);
-                                peg$currPos++;
-                            }
-                            else {
-                                s3 = peg$FAILED;
-                                if (peg$silentFails === 0) {
-                                    peg$fail(peg$c56);
-                                }
-                            }
-                            if (s3 === peg$FAILED) {
-                                if (peg$c57.test(input.charAt(peg$currPos))) {
-                                    s3 = input.charAt(peg$currPos);
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s3 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c58);
-                                    }
-                                }
-                                if (s3 === peg$FAILED) {
-                                    if (input.charCodeAt(peg$currPos) === 95) {
-                                        s3 = peg$c59;
-                                        peg$currPos++;
-                                    }
-                                    else {
-                                        s3 = peg$FAILED;
-                                        if (peg$silentFails === 0) {
-                                            peg$fail(peg$c60);
-                                        }
-                                    }
-                                    if (s3 === peg$FAILED) {
-                                        if (input.charCodeAt(peg$currPos) === 45) {
-                                            s3 = peg$c19;
-                                            peg$currPos++;
-                                        }
-                                        else {
-                                            s3 = peg$FAILED;
-                                            if (peg$silentFails === 0) {
-                                                peg$fail(peg$c20);
-                                            }
-                                        }
-                                        if (s3 === peg$FAILED) {
-                                            if (input.charCodeAt(peg$currPos) === 32) {
-                                                s3 = peg$c50;
-                                                peg$currPos++;
-                                            }
-                                            else {
-                                                s3 = peg$FAILED;
-                                                if (peg$silentFails === 0) {
-                                                    peg$fail(peg$c51);
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                else {
-                    s2 = peg$FAILED;
-                }
-                if (s2 !== peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 62) {
-                        s3 = peg$c64;
+                      if (input.charCodeAt(peg$currPos) === 32) {
+                        s3 = peg$c50;
                         peg$currPos++;
-                    }
-                    else {
+                      } else {
                         s3 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c65);
-                        }
+                        if (peg$silentFails === 0) { peg$fail(peg$c51); }
+                      }
                     }
-                    if (s3 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c66(s2);
-                        s0 = s1;
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
+                  }
                 }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
+              }
             }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            return s0;
+          }
+        } else {
+          s2 = peg$FAILED;
         }
-        function peg$parseBool() {
-            var s0, s1;
+        if (s2 !== peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 62) {
+            s3 = peg$c64;
+            peg$currPos++;
+          } else {
+            s3 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c65); }
+          }
+          if (s3 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c66(s2);
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+
+      return s0;
+    }
+
+    function peg$parseBool() {
+      var s0, s1;
+
+      s0 = peg$currPos;
+      if (input.charCodeAt(peg$currPos) === 48) {
+        s1 = peg$c67;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c68); }
+      }
+      if (s1 !== peg$FAILED) {
+        peg$savedPos = s0;
+        s1 = peg$c69();
+      }
+      s0 = s1;
+      if (s0 === peg$FAILED) {
+        s0 = peg$currPos;
+        if (input.charCodeAt(peg$currPos) === 49) {
+          s1 = peg$c70;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c71); }
+        }
+        if (s1 !== peg$FAILED) {
+          peg$savedPos = s0;
+          s1 = peg$c72();
+        }
+        s0 = s1;
+        if (s0 === peg$FAILED) {
+          s0 = peg$currPos;
+          if (input.substr(peg$currPos, 5) === peg$c73) {
+            s1 = peg$c73;
+            peg$currPos += 5;
+          } else {
+            s1 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c74); }
+          }
+          if (s1 !== peg$FAILED) {
+            peg$savedPos = s0;
+            s1 = peg$c69();
+          }
+          s0 = s1;
+          if (s0 === peg$FAILED) {
             s0 = peg$currPos;
-            if (input.charCodeAt(peg$currPos) === 48) {
-                s1 = peg$c67;
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c68);
-                }
+            if (input.substr(peg$currPos, 4) === peg$c75) {
+              s1 = peg$c75;
+              peg$currPos += 4;
+            } else {
+              s1 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c76); }
             }
             if (s1 !== peg$FAILED) {
-                peg$savedPos = s0;
-                s1 = peg$c69();
+              peg$savedPos = s0;
+              s1 = peg$c72();
             }
             s0 = s1;
             if (s0 === peg$FAILED) {
+              s0 = peg$currPos;
+              if (input.charCodeAt(peg$currPos) === 102) {
+                s1 = peg$c77;
+                peg$currPos++;
+              } else {
+                s1 = peg$FAILED;
+                if (peg$silentFails === 0) { peg$fail(peg$c78); }
+              }
+              if (s1 !== peg$FAILED) {
+                peg$savedPos = s0;
+                s1 = peg$c69();
+              }
+              s0 = s1;
+              if (s0 === peg$FAILED) {
                 s0 = peg$currPos;
-                if (input.charCodeAt(peg$currPos) === 49) {
-                    s1 = peg$c70;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c71);
-                    }
+                if (input.charCodeAt(peg$currPos) === 116) {
+                  s1 = peg$c79;
+                  peg$currPos++;
+                } else {
+                  s1 = peg$FAILED;
+                  if (peg$silentFails === 0) { peg$fail(peg$c80); }
                 }
                 if (s1 !== peg$FAILED) {
-                    peg$savedPos = s0;
-                    s1 = peg$c72();
+                  peg$savedPos = s0;
+                  s1 = peg$c72();
                 }
                 s0 = s1;
-                if (s0 === peg$FAILED) {
-                    s0 = peg$currPos;
-                    if (input.substr(peg$currPos, 5) === peg$c73) {
-                        s1 = peg$c73;
-                        peg$currPos += 5;
-                    }
-                    else {
-                        s1 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c74);
-                        }
-                    }
-                    if (s1 !== peg$FAILED) {
-                        peg$savedPos = s0;
-                        s1 = peg$c69();
-                    }
-                    s0 = s1;
-                    if (s0 === peg$FAILED) {
-                        s0 = peg$currPos;
-                        if (input.substr(peg$currPos, 4) === peg$c75) {
-                            s1 = peg$c75;
-                            peg$currPos += 4;
-                        }
-                        else {
-                            s1 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c76);
-                            }
-                        }
-                        if (s1 !== peg$FAILED) {
-                            peg$savedPos = s0;
-                            s1 = peg$c72();
-                        }
-                        s0 = s1;
-                        if (s0 === peg$FAILED) {
-                            s0 = peg$currPos;
-                            if (input.charCodeAt(peg$currPos) === 102) {
-                                s1 = peg$c77;
-                                peg$currPos++;
-                            }
-                            else {
-                                s1 = peg$FAILED;
-                                if (peg$silentFails === 0) {
-                                    peg$fail(peg$c78);
-                                }
-                            }
-                            if (s1 !== peg$FAILED) {
-                                peg$savedPos = s0;
-                                s1 = peg$c69();
-                            }
-                            s0 = s1;
-                            if (s0 === peg$FAILED) {
-                                s0 = peg$currPos;
-                                if (input.charCodeAt(peg$currPos) === 116) {
-                                    s1 = peg$c79;
-                                    peg$currPos++;
-                                }
-                                else {
-                                    s1 = peg$FAILED;
-                                    if (peg$silentFails === 0) {
-                                        peg$fail(peg$c80);
-                                    }
-                                }
-                                if (s1 !== peg$FAILED) {
-                                    peg$savedPos = s0;
-                                    s1 = peg$c72();
-                                }
-                                s0 = s1;
-                            }
-                        }
-                    }
-                }
+              }
             }
-            return s0;
+          }
         }
-        function peg$parseInteger() {
-            var s0, s1, s2, s3;
-            peg$silentFails++;
-            s0 = peg$currPos;
-            s1 = peg$parse_();
-            if (s1 !== peg$FAILED) {
-                s2 = [];
-                if (peg$c57.test(input.charAt(peg$currPos))) {
-                    s3 = input.charAt(peg$currPos);
-                    peg$currPos++;
-                }
-                else {
-                    s3 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c58);
-                    }
-                }
-                if (s3 !== peg$FAILED) {
-                    while (s3 !== peg$FAILED) {
-                        s2.push(s3);
-                        if (peg$c57.test(input.charAt(peg$currPos))) {
-                            s3 = input.charAt(peg$currPos);
-                            peg$currPos++;
-                        }
-                        else {
-                            s3 = peg$FAILED;
-                            if (peg$silentFails === 0) {
-                                peg$fail(peg$c58);
-                            }
-                        }
-                    }
-                }
-                else {
-                    s2 = peg$FAILED;
-                }
-                if (s2 !== peg$FAILED) {
-                    peg$savedPos = s0;
-                    s1 = peg$c82();
-                    s0 = s1;
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            peg$silentFails--;
-            if (s0 === peg$FAILED) {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c81);
-                }
-            }
-            return s0;
-        }
-        function peg$parseNxt() {
-            var s0;
-            s0 = peg$parseEOL();
-            if (s0 === peg$FAILED) {
-                s0 = peg$parseEOF();
-                if (s0 === peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 59) {
-                        s0 = peg$c83;
-                        peg$currPos++;
-                    }
-                    else {
-                        s0 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c84);
-                        }
-                    }
-                }
-            }
-            return s0;
-        }
-        function peg$parseEOL() {
-            var s0;
-            if (input.substr(peg$currPos, 2) === peg$c85) {
-                s0 = peg$c85;
-                peg$currPos += 2;
-            }
-            else {
-                s0 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c86);
-                }
-            }
-            if (s0 === peg$FAILED) {
-                if (input.charCodeAt(peg$currPos) === 10) {
-                    s0 = peg$c87;
-                    peg$currPos++;
-                }
-                else {
-                    s0 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c88);
-                    }
-                }
-            }
-            return s0;
-        }
-        function peg$parseEOF() {
-            var s0, s1;
-            s0 = peg$currPos;
-            peg$silentFails++;
-            if (input.length > peg$currPos) {
-                s1 = input.charAt(peg$currPos);
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c89);
-                }
-            }
-            peg$silentFails--;
-            if (s1 === peg$FAILED) {
-                s0 = void 0;
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            return s0;
-        }
-        function peg$parseSep() {
-            var s0, s1, s2, s3;
-            s0 = peg$currPos;
-            s1 = peg$parse_();
-            if (s1 !== peg$FAILED) {
-                if (input.charCodeAt(peg$currPos) === 44) {
-                    s2 = peg$c90;
-                    peg$currPos++;
-                }
-                else {
-                    s2 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c91);
-                    }
-                }
-                if (s2 !== peg$FAILED) {
-                    s3 = peg$parse_();
-                    if (s3 !== peg$FAILED) {
-                        s1 = [s1, s2, s3];
-                        s0 = s1;
-                    }
-                    else {
-                        peg$currPos = s0;
-                        s0 = peg$FAILED;
-                    }
-                }
-                else {
-                    peg$currPos = s0;
-                    s0 = peg$FAILED;
-                }
-            }
-            else {
-                peg$currPos = s0;
-                s0 = peg$FAILED;
-            }
-            return s0;
-        }
-        function peg$parse_() {
-            var s0, s1;
-            s0 = [];
-            if (input.charCodeAt(peg$currPos) === 32) {
-                s1 = peg$c50;
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c51);
-                }
-            }
-            if (s1 === peg$FAILED) {
-                if (input.charCodeAt(peg$currPos) === 9) {
-                    s1 = peg$c92;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c93);
-                    }
-                }
-            }
-            while (s1 !== peg$FAILED) {
-                s0.push(s1);
-                if (input.charCodeAt(peg$currPos) === 32) {
-                    s1 = peg$c50;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c51);
-                    }
-                }
-                if (s1 === peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 9) {
-                        s1 = peg$c92;
-                        peg$currPos++;
-                    }
-                    else {
-                        s1 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c93);
-                        }
-                    }
-                }
-            }
-            return s0;
-        }
-        function peg$parse__() {
-            var s0, s1;
-            s0 = [];
-            if (input.charCodeAt(peg$currPos) === 32) {
-                s1 = peg$c50;
-                peg$currPos++;
-            }
-            else {
-                s1 = peg$FAILED;
-                if (peg$silentFails === 0) {
-                    peg$fail(peg$c51);
-                }
-            }
-            if (s1 === peg$FAILED) {
-                if (input.charCodeAt(peg$currPos) === 9) {
-                    s1 = peg$c92;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c93);
-                    }
-                }
-                if (s1 === peg$FAILED) {
-                    s1 = peg$parseEOL();
-                }
-            }
-            while (s1 !== peg$FAILED) {
-                s0.push(s1);
-                if (input.charCodeAt(peg$currPos) === 32) {
-                    s1 = peg$c50;
-                    peg$currPos++;
-                }
-                else {
-                    s1 = peg$FAILED;
-                    if (peg$silentFails === 0) {
-                        peg$fail(peg$c51);
-                    }
-                }
-                if (s1 === peg$FAILED) {
-                    if (input.charCodeAt(peg$currPos) === 9) {
-                        s1 = peg$c92;
-                        peg$currPos++;
-                    }
-                    else {
-                        s1 = peg$FAILED;
-                        if (peg$silentFails === 0) {
-                            peg$fail(peg$c93);
-                        }
-                    }
-                    if (s1 === peg$FAILED) {
-                        s1 = peg$parseEOL();
-                    }
-                }
-            }
-            return s0;
-        }
-        function peg$parseEmpty() {
-            var s0;
-            s0 = peg$c46;
-            return s0;
-        }
-        var graphs = [];
-        var top = function () { return graphs[graphs.length - 1]; };
-        peg$result = peg$startRuleFunction();
-        if (peg$result !== peg$FAILED && peg$currPos === input.length) {
-            return peg$result;
-        }
-        else {
-            if (peg$result !== peg$FAILED && peg$currPos < input.length) {
-                peg$fail(peg$endExpectation());
-            }
-            throw peg$buildStructuredError(peg$maxFailExpected, peg$maxFailPos < input.length ? input.charAt(peg$maxFailPos) : null, peg$maxFailPos < input.length
-                ? peg$computeLocation(peg$maxFailPos, peg$maxFailPos + 1)
-                : peg$computeLocation(peg$maxFailPos, peg$maxFailPos));
-        }
+      }
+
+      return s0;
     }
-    return {
-        SyntaxError: peg$SyntaxError,
-        parse: peg$parse
-    };
+
+    function peg$parseInteger() {
+      var s0, s1, s2, s3;
+
+      peg$silentFails++;
+      s0 = peg$currPos;
+      s1 = peg$parse_();
+      if (s1 !== peg$FAILED) {
+        s2 = [];
+        if (peg$c57.test(input.charAt(peg$currPos))) {
+          s3 = input.charAt(peg$currPos);
+          peg$currPos++;
+        } else {
+          s3 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c58); }
+        }
+        if (s3 !== peg$FAILED) {
+          while (s3 !== peg$FAILED) {
+            s2.push(s3);
+            if (peg$c57.test(input.charAt(peg$currPos))) {
+              s3 = input.charAt(peg$currPos);
+              peg$currPos++;
+            } else {
+              s3 = peg$FAILED;
+              if (peg$silentFails === 0) { peg$fail(peg$c58); }
+            }
+          }
+        } else {
+          s2 = peg$FAILED;
+        }
+        if (s2 !== peg$FAILED) {
+          peg$savedPos = s0;
+          s1 = peg$c82();
+          s0 = s1;
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+      peg$silentFails--;
+      if (s0 === peg$FAILED) {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c81); }
+      }
+
+      return s0;
+    }
+
+    function peg$parseNxt() {
+      var s0;
+
+      s0 = peg$parseEOL();
+      if (s0 === peg$FAILED) {
+        s0 = peg$parseEOF();
+        if (s0 === peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 59) {
+            s0 = peg$c83;
+            peg$currPos++;
+          } else {
+            s0 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c84); }
+          }
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseEOL() {
+      var s0;
+
+      if (input.substr(peg$currPos, 2) === peg$c85) {
+        s0 = peg$c85;
+        peg$currPos += 2;
+      } else {
+        s0 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c86); }
+      }
+      if (s0 === peg$FAILED) {
+        if (input.charCodeAt(peg$currPos) === 10) {
+          s0 = peg$c87;
+          peg$currPos++;
+        } else {
+          s0 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c88); }
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseEOF() {
+      var s0, s1;
+
+      s0 = peg$currPos;
+      peg$silentFails++;
+      if (input.length > peg$currPos) {
+        s1 = input.charAt(peg$currPos);
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c89); }
+      }
+      peg$silentFails--;
+      if (s1 === peg$FAILED) {
+        s0 = void 0;
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+
+      return s0;
+    }
+
+    function peg$parseSep() {
+      var s0, s1, s2, s3;
+
+      s0 = peg$currPos;
+      s1 = peg$parse_();
+      if (s1 !== peg$FAILED) {
+        if (input.charCodeAt(peg$currPos) === 44) {
+          s2 = peg$c90;
+          peg$currPos++;
+        } else {
+          s2 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c91); }
+        }
+        if (s2 !== peg$FAILED) {
+          s3 = peg$parse_();
+          if (s3 !== peg$FAILED) {
+            s1 = [s1, s2, s3];
+            s0 = s1;
+          } else {
+            peg$currPos = s0;
+            s0 = peg$FAILED;
+          }
+        } else {
+          peg$currPos = s0;
+          s0 = peg$FAILED;
+        }
+      } else {
+        peg$currPos = s0;
+        s0 = peg$FAILED;
+      }
+
+      return s0;
+    }
+
+    function peg$parse_() {
+      var s0, s1;
+
+      s0 = [];
+      if (input.charCodeAt(peg$currPos) === 32) {
+        s1 = peg$c50;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c51); }
+      }
+      if (s1 === peg$FAILED) {
+        if (input.charCodeAt(peg$currPos) === 9) {
+          s1 = peg$c92;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c93); }
+        }
+      }
+      while (s1 !== peg$FAILED) {
+        s0.push(s1);
+        if (input.charCodeAt(peg$currPos) === 32) {
+          s1 = peg$c50;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c51); }
+        }
+        if (s1 === peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 9) {
+            s1 = peg$c92;
+            peg$currPos++;
+          } else {
+            s1 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c93); }
+          }
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parse__() {
+      var s0, s1;
+
+      s0 = [];
+      if (input.charCodeAt(peg$currPos) === 32) {
+        s1 = peg$c50;
+        peg$currPos++;
+      } else {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c51); }
+      }
+      if (s1 === peg$FAILED) {
+        if (input.charCodeAt(peg$currPos) === 9) {
+          s1 = peg$c92;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c93); }
+        }
+        if (s1 === peg$FAILED) {
+          s1 = peg$parseEOL();
+        }
+      }
+      while (s1 !== peg$FAILED) {
+        s0.push(s1);
+        if (input.charCodeAt(peg$currPos) === 32) {
+          s1 = peg$c50;
+          peg$currPos++;
+        } else {
+          s1 = peg$FAILED;
+          if (peg$silentFails === 0) { peg$fail(peg$c51); }
+        }
+        if (s1 === peg$FAILED) {
+          if (input.charCodeAt(peg$currPos) === 9) {
+            s1 = peg$c92;
+            peg$currPos++;
+          } else {
+            s1 = peg$FAILED;
+            if (peg$silentFails === 0) { peg$fail(peg$c93); }
+          }
+          if (s1 === peg$FAILED) {
+            s1 = peg$parseEOL();
+          }
+        }
+      }
+
+      return s0;
+    }
+
+    function peg$parseEmpty() {
+      var s0;
+
+      s0 = peg$c46;
+
+      return s0;
+    }
+
+
+      var graphs = [];
+      var top = function() {return graphs[graphs.length-1];}    
+
+
+    peg$result = peg$startRuleFunction();
+
+    if (peg$result !== peg$FAILED && peg$currPos === input.length) {
+      return peg$result;
+    } else {
+      if (peg$result !== peg$FAILED && peg$currPos < input.length) {
+        peg$fail(peg$endExpectation());
+      }
+
+      throw peg$buildStructuredError(
+        peg$maxFailExpected,
+        peg$maxFailPos < input.length ? input.charAt(peg$maxFailPos) : null,
+        peg$maxFailPos < input.length
+          ? peg$computeLocation(peg$maxFailPos, peg$maxFailPos + 1)
+          : peg$computeLocation(peg$maxFailPos, peg$maxFailPos)
+      );
+    }
+  }
+
+  return {
+    SyntaxError: peg$SyntaxError,
+    parse:       peg$parse
+  };
 })();
-exports.parser = parser;
